Backend presque terminé avec le cours (API Inaccessible)

+++ back/controllers/sauce.js
@@ -1,26 +1,57 @@
 const sauce = require('../models/sauce');
+const fs = require('fs');
 
 
 exports.createSauce = (req, res, next) => {
-    delete req.body._id;
-    const sauce = new Sauce({
-      ...req.body
-    });
+    const sauceObject = JSON.parse(req.body.sauce); 
+    delete sauceObject._id;
+    delete sauceObject._userId;
+    const sauce = new sauce({
+        ...sauceObject,
+        userId: req.auth.userId,
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    })
     sauce.save()
-    .then(() => res.status(201).json({message: 'Sauce ajoutée !'}))
-    .catch(error => res.status(400).json({error}));
+    .then(() => {res.status(201).json({message: 'Sauce ajoutée !'})})
+    .catch(error => {res.status(400).json({error})});
 };
 
 exports.modifySauce = (req, res, next) => {
-    sauce.updateOne({_id: req.params.id}, {...req.body, _id:req.params.id})
-    .then(() => res.status(200).json({message: 'Sauce modifiée !'}))
-    .catch(error => res.status(400).json({error}));
+    const sauceObject = req.file ?{
+        ...JSON.parse(req.body.sauce),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : {...req.body};
+
+    delete sauceObject._userId;
+
+    sauce.findOne({id: req.params.id})
+        .then(sauce => {
+            if(sauce.userId != req.userId){
+                res.status(401).json({message: 'Non-autorisé'})
+            }else{
+                sauce.updateOne({_id: req.params.id}, {...sauceObject, _id: req.params.id})
+                .then(() => res.status(200).json({message: 'Sauce modifiée'}))
+                .catch(error => res.status(401).json({error}));
+            }
+        })
+        .catch(error => {res.status(400).json({error})});
 };
 
 exports.deleteSauce = (req, res, next) => {
-    sauce.deleteOne({_id: req.params.id})
-    .then(() => res.status(200).json({message: 'Sauce supprimée !'}))
-    .catch(error => res.status(400).json({error}));
+    sauce.findOne({id: req.params.id})
+        .then(sauce => {
+            if(sauce.userId != req.auth.userId){
+                res.status(401).json({message: 'Non-autorisé'});
+            }else{
+                const filename = sauce.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    sauce.deleteOne({_id: req.params.id})
+                        .then(() => {res.status(200).json({message: 'Sauce supprimée !'})})
+                        .catch(error => res.status(401).json({error}));
+                })
+            }
+        })
+        .catch(error => {res.status(500).json({error})});
 };
 
 exports.getOneSauce = (req, res, next) => {
